# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controller.SuccessResponse:
    properties:
      result:
        type: string
    type: object
...
  domain.CreateHackathonRequestBody:
    properties:
      description:
        type: string
      expired:
        type: string
      link:
        type: string
      name:
        type: string
      start_date:
        type: string
      status_tags:
        items:
          type: integer
        type: array
      term:
        type: integer
    type: object
...
  domain.HackathonResponses:
    properties:
      description:
        type: string
      expired:
        type: string
      hackathon_id:
        type: integer
      icon:
        type: string
      link:
        type: string
      name:
        type: string
      start_date:
        type: string
      status_tags:
        items:
          $ref: '#/definitions/repository.StatusTag'
        type: array
      term:
        type: integer
    type: object
  domain.ListPastWorkResponse:
    properties:
      explanatory_text:
        type: string
      frameworks:
        items:
          $ref: '#/definitions/repository.Framework'
        type: array
      members:
        items:
          $ref: '#/definitions/domain.PastWorkMembers'
        type: array
      name:
        type: string
      opus:
        type: integer
      tech_tags:
        items:
          $ref: '#/definitions/repository.TechTag'
        type: array
    type: object
...
  domain.PastWorkMembers:
    properties:
      account_id:
        type: string
      icon:
        type: string
      name:
        type: string
    type: object
  domain.PastWorkRequestBody:
    properties:
      account_past_works:
        type: string
      explanatory_text:
        type: string
      name:
        type: string
      past_work_frameworks:
        type: string
      past_work_tags:
        type: string
    type: object
  domain.PastWorkResponse:
    properties:
      award_data_id:
        type: integer
      create_at:
        type: string
      explanatory_text:
        type: string
      frameworks:
        items:
          $ref: '#/definitions/repository.Framework'
        type: array
      is_delete:
        type: boolean
      members:
        items:
          $ref: '#/definitions/domain.PastWorkMembers'
        type: array
      name:
        type: string
      opus:
        type: integer
      tech_tags:
        items:
          $ref: '#/definitions/repository.TechTag'
        type: array
      thumbnail_image:
        type: string
      update_at:
        type: string
    type: object
...
  repository.Framework:
    properties:
      framework:
        type: string
      framework_id:
        type: integer
      tech_tag_id:
        type: integer
    type: object
...
  repository.StatusTag:
    properties:
      status:
        type: string
      status_id:
        type: integer
    type: object
  repository.TechTag:
    properties:
      icon:
        type: string
      language:
        type: string
      tech_tag_id:
        type: integer
    type: object
host: https://seaffood.com
info:
  contact:
    email: oogiriminister@gmail.com
    name: murasame
    url: https://twitter.com/fresh_salmon256
  description: HackPortal Backend API serice
  license:
    name: No-license
    url: No-license
  termsOfService: https://seaffood.com/api
  title: Hack Hack Backend API
  version: "1.0"
paths:
...
  /hackathons:
    get:
      description: List Hackathon
      parameters:
      - in: query
        name: expired
        type: boolean
      - in: query
        name: page_id
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/domain.HackathonResponses'
            type: array
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List Hackathon
      tags:
      - Hackathon
    post:
      description: Register a hackathon from given parameters
      parameters:
      - description: create hackathon Request Body
        in: body
        name: CreateHackathonRequestBody
        required: true
        schema:
          $ref: '#/definitions/domain.CreateHackathonRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/domain.HackathonResponses'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create Hackathon
      tags:
      - Hackathon
  /hackathons/{hackathon_id}:
    get:
      description: Get Hackathon
      parameters:
      - description: Hackathons API wildcard
        in: path
        name: hackathon_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/domain.HackathonResponses'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get Hackathon
      tags:
      - Hackathon
...
  /pastworks:
    get:
      description: List PastWork
      parameters:
      - in: query
        name: page_id
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ListPastWorkResponse'
            type: array
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List PastWork
      tags:
      - PastWorks
    post:
      description: Create a past work from the requested body
      parameters:
      - description: Create PastWork Request
        in: body
        name: CreatePastWorkRequest
        required: true
        schema:
          $ref: '#/definitions/domain.PastWorkRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: create success response
          schema:
            $ref: '#/definitions/domain.PastWorkResponse'
        "400":
          description: bad request response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: server error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create new past work
      tags:
      - PastWorks
  /pastworks/{opus}:
    delete:
      description: Delete PastWork
      parameters:
      - description: PastWorks API wildcard
        in: path
        name: opus
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Delete PastWork
      tags:
      - PastWorks
    get:
      description: Get PastWork
      parameters:
      - description: PastWorks API wildcard
        in: path
        name: opus
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PastWorkResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get PastWork
      tags:
      - PastWorks
    put:
      description: Update PastWork
      parameters:
      - description: PastWorks API wildcard
        in: path
        name: opus
        required: true
        type: string
      - description: Update PastWork Request
        in: body
        name: UpdatePastWorkRequest
        required: true
        schema:
          $ref: '#/definitions/domain.PastWorkRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PastWorkResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Update PastWork
      tags:
      - PastWorks
...
swagger: "2.0"